refactor: 필요없는 패키지 제거 후 통합
