+++ src/main/java/com/softeer/podoarrival/event/model/entity/Role.java
@@ -1,4 +1,4 @@
-package com.softeer.podoarrival.user.model.entity;
+package com.softeer.podoarrival.event.model.entity;
 
 public enum Role {
     ROLE_USER, ROLE_ADMIN, ROLE_ANONYMOUS
