+++ src/main/java/com/softeer/podoarrival/event/repository/ArrivalUserRepository.java
@@ -1,0 +1,7 @@
+package com.softeer.podoarrival.event.repository;
+
+import com.softeer.podoarrival.event.model.entity.ArrivalUser;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ArrivalUserRepository extends JpaRepository<ArrivalUser, Long> {
+}
