+++ src/main/java/com/softeer/podoarrival/event/service/ArrivalEventReleaseService.java
@@ -3,9 +3,9 @@ package com.softeer.podoarrival.event.service;
 import com.softeer.podoarrival.event.exception.ExistingUserException;
 import com.softeer.podoarrival.event.model.dto.ArrivalApplicationResponseDto;
 import com.softeer.podoarrival.security.AuthInfo;
-import com.softeer.podoarrival.user.model.entity.ArrivalUser;
-import com.softeer.podoarrival.user.model.entity.Role;
-import com.softeer.podoarrival.user.repository.ArrivalUserRepository;
+import com.softeer.podoarrival.event.model.entity.ArrivalUser;
+import com.softeer.podoarrival.event.model.entity.Role;
+import com.softeer.podoarrival.event.repository.ArrivalUserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.redisson.api.BatchResult;
